Refactored code to allow easier integration of kernel rendering
Renderer.hpp and Renderer.cpp should now contain most of the agnostic code, alowing for easier code reuse
Some renaming may still be required
